Fix -Wimplicit-int-float-conversions.
Bug: chromium:989932
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/2987733
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>

+++ src/libANGLE/renderer/copyvertex.inc.h
@@ -204,12 +204,12 @@ inline void CopyToFloatVertexData(const uint8_t *input,
             {
                 if (NL::is_signed)
                 {
-                    result = static_cast<float>(offsetInput[j]) / NL::max();
+                    result = static_cast<float>(offsetInput[j]) / static_cast<float>(NL::max());
                     result = result >= -1.0f ? result : -1.0f;
                 }
                 else
                 {
-                    result = static_cast<float>(offsetInput[j]) / NL::max();
+                    result = static_cast<float>(offsetInput[j]) / static_cast<float>(NL::max());
                 }
             }
             else
